Update to latest OpenBSD code.
Brings in these two commits:
---
date: 2020/05/22 07:50:07;  author: benno;  state: Exp;  lines: +7 -1;  commitid: H8MJhjoH6o7bDIkz;
Fix the exit code when eval()uating a || compound list, it would
terminate the shell when running under -e.
See also https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=269067 and
Bug reported including fix by Leah Neukirchen, Thanks!
ok millert@
---

---
date: 2020/07/07 10:33:58;  author: jca;  state: Exp;  lines: +28 -2;  commitid: Yo5zapKJfyJKPH39;
Add support for set -o pipefail

With the pipefail option set, the exit status of a pipeline is 0 if all
commands succeed, or the return status of the rightmost command that
fails.  This can help stronger error checking, but is not a silver
bullet.  For example, commands will exhibit a non-zero exit status if
they're killed by a SIGPIPE when writing to a pipe.  Yet pipefail was
considered useful enough to be included in the next POSIX standard.

This implementation remembers the value of the pipefail option when
a pipeline is started, as described as option 1) in

  https://www.austingroupbugs.net/view.php?id=789#c4102

Requested by ajacoutot@, ok millert@
---