Make stripping of executables optional
Some operating systems and distributions, like Fedora, use the debug
information of the installed executables to generate packages that may
be used by the user and the distribution packages to debug problems
with the installation.

This patch makes it so that a builder may optionally disable the
stripping of executables at ./configure-time.